Move the base manifest and package reset to platform/base_manifest.rs
The safety critical delete logic and its tests now live side by side.
python_tree_root and BASE_MANIFEST become pub(super) for callers elsewhere
in platform; the reset entry points are re-exported so call sites are
untouched. can_refresh_from_bundle moves too, so #335 can retire the whole
subsystem by deleting one file.
